refactor: 주소에 맞게 페이지를 수정합니다.
