--- app/(app)/project/[id]/review/delete.tsx
@@ -1,13 +1,0 @@
-import { View } from 'react-native';
-
-import Typography from '@/components/common/typography';
-
-function Delete() {
-  return (
-    <View>
-      <Typography>Delete</Typography>
-    </View>
-  );
-}
-
-export default Delete;
