+++ src/constants/siteUrls.ts
@@ -9,4 +9,5 @@ export const PROJECT_URLS = {
   PROJECT_CREATE: '/project/create',
   PROJECT_DETAIL: '/project/[id]/detail',
   PROJECT_REVIEW: '/project/[id]/review',
+  PROJECT_REVIEW_CREATE: '/project/[id]/review/create',
 } as const;
